Moved token definitions into utils
